requirements.txt
streamlit
embedchain
streamlit-chat
temp
ollama
langchain_community
requests
PyPDF2
chromadb
streamlit-extras